+++ test/ComposerRequireCheckerTest/ASTLocator/LocateASTFromFilesTest.php
@@ -6,6 +6,8 @@ use ArrayObject;
 use ComposerRequireChecker\ASTLocator\LocateASTFromFiles;
 use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamDirectory;
+use PhpParser\Error;
+use PhpParser\ErrorHandler\Collecting;
 use PhpParser\Lexer;
 use PhpParser\Parser\Php7;
 use PHPUnit\Framework\TestCase;
@@ -24,7 +26,7 @@ class LocateASTFromFilesTest extends TestCase
     {
         parent::setUp();
 
-        $this->locator = new LocateASTFromFiles(new Php7(new Lexer()));
+        $this->locator = new LocateASTFromFiles(new Php7(new Lexer()), null);
         $this->root = vfsStream::setup();
     }
 
@@ -38,6 +40,30 @@ class LocateASTFromFilesTest extends TestCase
         $roots = $this->locate($files);
 
         $this->assertCount(2, $roots);
+    }
+
+    public function testFailOnParseError()
+    {
+        $this->expectException(Error::class);
+        $files = [
+            $this->createFile('MyBadCode', '<?php this causes a parse error'),
+        ];
+
+        $this->locate($files);
+    }
+
+    public function testDoNotFailOnParseErrorWithErrorHandler()
+    {
+        $collectingErrorHandler = new Collecting();
+        $this->locator = new LocateASTFromFiles(new Php7(new Lexer()), $collectingErrorHandler);
+        $files = [
+            $this->createFile('MyBadCode', '<?php this causes a parse error'),
+        ];
+
+        $roots = $this->locate($files);
+        $this->assertCount(1, $roots); // one file should be parsed (partially)
+        $this->assertTrue($collectingErrorHandler->hasErrors());
+        $this->assertCount(1, $collectingErrorHandler->getErrors()); //should have one parse error
     }
 
     private function createFile(string $path, string $content = null): string