+++ test/ComposerRequireCheckerTest/DefinedSymbolsLocator/LocateDefinedSymbolsFromASTRootsTest.php
@@ -5,6 +5,8 @@ namespace ComposerRequireCheckerTest\DefinedSymbolsLocator;
 use ArrayObject;
 use ComposerRequireChecker\DefinedSymbolsLocator\LocateDefinedSymbolsFromASTRoots;
 use PhpParser\Node\Stmt\Class_;
+use PhpParser\Node\Stmt\Function_;
+use PhpParser\Node\Stmt\Trait_;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -29,7 +31,7 @@ class LocateDefinedSymbolsFromASTRootsTest extends TestCase
         $this->assertCount(0, $symbols);
     }
 
-    public function testBasicLocate()
+    public function testBasicLocateClass()
     {
         $roots = [
             [new Class_('MyClassA'), new Class_('MyClassB')],
@@ -44,6 +46,39 @@ class LocateDefinedSymbolsFromASTRootsTest extends TestCase
         $this->assertContains('MyClassA', $symbols);
         $this->assertContains('MyClassB', $symbols);
         $this->assertContains('MyClassC', $symbols);
+    }
+
+    public function testBasicLocateFunctions()
+    {
+        $roots = [
+            [new Function_('myFunctionA')],
+            [new Class_('myFunctionB')],
+        ];
+
+        $symbols = $this->locate([$roots]);
+
+        $this->assertInternalType('array', $symbols);
+        $this->assertCount(2, $symbols);
+
+        $this->assertContains('myFunctionA', $symbols);
+        $this->assertContains('myFunctionB', $symbols);
+    }
+
+    public function testBasicLocateTrait()
+    {
+        $roots = [
+            [new Trait_('MyTraitA'), new Trait_('MyTraitB')],
+            [new Trait_('MyTraitC')],
+        ];
+
+        $symbols = $this->locate([$roots]);
+
+        $this->assertInternalType('array', $symbols);
+        $this->assertCount(3, $symbols);
+
+        $this->assertContains('MyTraitA', $symbols);
+        $this->assertContains('MyTraitB', $symbols);
+        $this->assertContains('MyTraitC', $symbols);
     }
 
     public function testBasicLocateAnonymous()