+++ CHANGELOG.md
@@ -8,9 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- PHP 7.1 required
 - displays more detailed errors if json config format is not valid
 - fixed fatal error when parsing trait usage with modified visibility (#44)
 - when locating files "exclude-from-classmap" will be taken into account
+- add option "--fail-on-parse-error"  
+  ComposerRequireChecker will only throw parser exceptions if this is set (#35)
+- add option "--ignore-parse-errors"  
+- add option "--ignore-parse-errors"
+  ComposerRequireChecker will not throw parser exceptions if this is set (#35)
 
 ## [0.1.6] - 2017-09-24
 ### Added